Internalize things that shouldn't be public
Originally Future class was intended to be public, but since this project adopted Swift-NIO, local Future class became just a proxy between FoundationDB and NIO's EventLoopFutures, and therefore should be internal now.

Major version bump is not needed, because module users couldn't interact with it anyway even though it was public.