Make `FDBTuplePackable` methods more specific (and less general). More.
Currently this protocol requires a very generic method name `get` (and
`_get`) which is not good for readability, clarity and generally not
swifty. Those methods should be renamed to be more FDB-specific.

Additional formatting cleanup should be done. And maybe the whole thing
will be a few CPU ticks faster because of more inlinable methods.

Closes #55.